validator when the file uploaded and validation failed delete the uploaded file

+++ src/middlewares/validator.js
@@ -11,9 +11,12 @@ const validator = (schema) => {
       });
       next();
     } catch (err) {
-      fs.unlink(req.file.path, (e) => {
-        if (e) console.log(e);
-      });
+      if (req.file) {
+        fs.unlink(req.file.path, (e) => {
+          if (e) console.log(e);
+        });
+      }
+
       return res.status(500).json({ message: err.errors });
     }
   };
